[CIT-309] move the thin edge json to separate module
